Update and rename README.md to Legend Jeep-Best Jeep Dealership in Kanpur.md
Don't delay any further! Initiate contact with Legend Jeep's authorized Jeep dealership in Kanpur today. Unearth the mysteries of the Jeep Compass's price, explore the spectrum of model choices, and pave the way for a Jeep adventure that echoes with excitement. With Legend Jeep, each ride transforms into an expedition of exhilaration and liberation, driven by the indomitable spirit of the iconic and versatile Jeep Compass.